requirements.txt
# Discord Bot Framework
discord.py>=2.3.2

# OpenAI Integration
openai>=1.3.0

# Anthropic Integration
anthropic>=0.39.0

# Environment Variables
python-dotenv>=1.0.0

# Async HTTP requests
aiohttp>=3.9.0

# Logging and utilities
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

chromadb>=0.4.15
sentence-transformers>=2.2.2

# Code formatting
black>=23.0.0